requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
python-dotenv
stripe
requests
